Move ArchivePasswordCache to file_operations module
