+++ file_operations/build.gradle
@@ -53,6 +53,18 @@ ext {
 dependencies {
     implementation 'org.apache.commons:commons-compress:1.18'
 
+    implementation ('com.github.junrar:junrar:0.7') {
+        exclude module: 'commons-logging'
+        exclude module: 'commons-logging-api'
+        exclude module: 'maven-scm-api'
+        exclude module: 'maven-scm-provider-svnexe'
+    }
+
+    //zip4j: support password-protected zips
+    implementation 'net.lingala.zip4j:zip4j:1.3.2'
+
+    implementation project(':commons_compress_7z')
+
     //SFTP
     implementation "com.hierynomus:sshj:$sshjVersion"
